move compute_d2dexdz2_fac_mc to reference_atmosphere.py
