+++ model/common/src/icon4py/model/common/metrics/reference_atmosphere.py
@@ -7,10 +7,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 import gt4py.next as gtx
 from gt4py.next import GridType, field_operator, program
-from gt4py.next.ffront.fbuiltins import exp, log
+from gt4py.next.ffront.fbuiltins import astype, exp, log
 
 from icon4py.model.common import dimension as dims, field_type_aliases as fa
-from icon4py.model.common.type_alias import wpfloat
+from icon4py.model.common.type_alias import vpfloat, wpfloat
 
 
 @field_operator
@@ -210,3 +210,87 @@ def compute_d_exner_dz_ref_ic(
     Returns: first vertical derivative of reference exner pressure
     """
     return -grav / (cpd * theta_ref_ic)
+
+
+@field_operator
+def _compute_d2dexdz2_fac_mc(
+    theta_ref_mc: fa.CellKField[vpfloat],
+    inv_ddqz_z_full: fa.CellKField[vpfloat],
+    exner_ref_mc: fa.CellKField[vpfloat],
+    z_mc: fa.CellKField[wpfloat],
+    cpd: wpfloat,
+    grav: wpfloat,
+    del_t_bg: wpfloat,
+    h_scal_bg: wpfloat,
+) -> tuple[fa.CellKField[vpfloat], fa.CellKField[vpfloat]]:
+    del_t_bg = astype(del_t_bg, vpfloat)
+    cpd = astype(cpd, vpfloat)
+    grav = astype(grav, vpfloat)
+    h_scal_bg = astype(h_scal_bg, vpfloat)
+    z_mc = astype(z_mc, vpfloat)
+    fac1 = -grav / (cpd * theta_ref_mc**2) * inv_ddqz_z_full
+    fac2 = (
+        2.0
+        * grav
+        / (cpd * theta_ref_mc**3)
+        * (grav / cpd - del_t_bg / h_scal_bg * exp(-z_mc / h_scal_bg))
+        / exner_ref_mc
+    )
+
+    return fac1, fac2
+
+
+@program(grid_type=GridType.UNSTRUCTURED)
+def compute_d2dexdz2_fac_mc(
+    theta_ref_mc: fa.CellKField[vpfloat],
+    inv_ddqz_z_full: fa.CellKField[vpfloat],
+    exner_ref_mc: fa.CellKField[vpfloat],
+    z_mc: fa.CellKField[wpfloat],
+    d2dexdz2_fac1_mc: fa.CellKField[vpfloat],
+    d2dexdz2_fac2_mc: fa.CellKField[vpfloat],
+    cpd: float,
+    grav: wpfloat,
+    del_t_bg: wpfloat,
+    h_scal_bg: wpfloat,
+    horizontal_start: gtx.int32,
+    horizontal_end: gtx.int32,
+    vertical_start: gtx.int32,
+    vertical_end: gtx.int32,
+):
+    """
+    Compute d2dexdz2_fac1_mc and d2dexdz2_fac2_mc factors.
+
+    See mo_vertical_grid.f90
+
+    Args:
+        theta_ref_mc: reference Potential temperature, full level mass points
+        inv_ddqz_z_full: inverse layer thickness (for runtime optimization)
+        exner_ref_mc: reference Exner pressure, full level mass points
+        z_mc: geometric height defined on full levels
+        d2dexdz2_fac1_mc: (output) first vertical derivative of reference Exner pressure, full level mass points, divided by theta_ref
+        d2dexdz2_fac2_mc: (output) vertical derivative of d_exner_dz/theta_ref, full level mass points
+        cpd: Specific heat at constant pressure [J/K/kg]
+        grav: average gravitational acceleration
+        del_t_bg: difference between sea level temperature and asymptotic stratospheric temperature
+        h_scal_bg: height scale for reference atmosphere [m]
+        horizontal_start: horizontal start index
+        horizontal_end: horizontal end index
+        vertical_start: vertical start index
+        vertical_end: vertical end index
+    """
+
+    _compute_d2dexdz2_fac_mc(
+        theta_ref_mc=theta_ref_mc,
+        inv_ddqz_z_full=inv_ddqz_z_full,
+        exner_ref_mc=exner_ref_mc,
+        z_mc=z_mc,
+        cpd=cpd,
+        grav=grav,
+        del_t_bg=del_t_bg,
+        h_scal_bg=h_scal_bg,
+        out=(d2dexdz2_fac1_mc, d2dexdz2_fac2_mc),
+        domain={
+            dims.CellDim: (horizontal_start, horizontal_end),
+            dims.KDim: (vertical_start, vertical_end),
+        },
+    )
