+++ model/common/tests/metric_tests/test_reference_atmosphere.py
@@ -17,6 +17,7 @@ from icon4py.model.common.interpolation.stencils.cell_2_edge_interpolation impor
 )
 from icon4py.model.common.metrics.metric_fields import compute_z_mc
 from icon4py.model.common.metrics.reference_atmosphere import (
+    compute_d2dexdz2_fac_mc,
     compute_d_exner_dz_ref_ic,
     compute_reference_atmosphere_cell_fields,
     compute_reference_atmosphere_edge_fields,
@@ -87,7 +88,7 @@ def test_compute_reference_atmosphere_fields_on_full_level_masspoints(
 
 @pytest.mark.datatest
 @pytest.mark.parametrize("experiment", [dt_utils.REGIONAL_EXPERIMENT, dt_utils.GLOBAL_EXPERIMENT])
-def test_compute_reference_atmsophere_on_half_level_mass_points(
+def test_compute_reference_atmosphere_on_half_level_mass_points(
     icon_grid, metrics_savepoint, backend
 ):
     theta_ref_ic_ref = metrics_savepoint.theta_ref_ic()
@@ -213,3 +214,35 @@ def test_compute_reference_atmosphere_on_full_level_edge_fields(
     )
     assert helpers.dallclose(rho_ref_me.asnumpy(), rho_ref_me_ref.asnumpy())
     assert helpers.dallclose(theta_ref_me.asnumpy(), theta_ref_me_ref.asnumpy())
+
+
+@pytest.mark.datatest
+@pytest.mark.parametrize("experiment", [dt_utils.REGIONAL_EXPERIMENT, dt_utils.GLOBAL_EXPERIMENT])
+def test_compute_d2dexdz2_fac_mc(icon_grid, metrics_savepoint, grid_savepoint, backend):
+    z_mc = metrics_savepoint.z_mc()
+    d2dexdz2_fac1_mc_ref = metrics_savepoint.d2dexdz2_fac1_mc()
+    d2dexdz2_fac2_mc_ref = metrics_savepoint.d2dexdz2_fac2_mc()
+
+    d2dexdz2_fac1_mc = data_alloc.zero_field(icon_grid, dims.CellDim, dims.KDim)
+    d2dexdz2_fac2_mc = data_alloc.zero_field(icon_grid, dims.CellDim, dims.KDim)
+
+    compute_d2dexdz2_fac_mc.with_backend(backend=backend)(
+        theta_ref_mc=metrics_savepoint.theta_ref_mc(),
+        inv_ddqz_z_full=metrics_savepoint.inv_ddqz_z_full(),
+        exner_ref_mc=metrics_savepoint.exner_ref_mc(),
+        z_mc=z_mc,
+        d2dexdz2_fac1_mc=d2dexdz2_fac1_mc,
+        d2dexdz2_fac2_mc=d2dexdz2_fac2_mc,
+        cpd=constants.CPD,
+        grav=constants.GRAV,
+        del_t_bg=constants.DEL_T_BG,
+        h_scal_bg=constants.HEIGHT_SCALE_FOR_REFERENCE_ATMOSPHERE,
+        horizontal_start=gtx.int32(0),
+        horizontal_end=gtx.int32(icon_grid.num_cells),
+        vertical_start=gtx.int32(0),
+        vertical_end=gtx.int32(icon_grid.num_levels),
+        offset_provider={},
+    )
+
+    assert helpers.dallclose(d2dexdz2_fac1_mc.asnumpy(), d2dexdz2_fac1_mc_ref.asnumpy())
+    assert helpers.dallclose(d2dexdz2_fac2_mc.asnumpy(), d2dexdz2_fac2_mc_ref.asnumpy())
